+++ MSAL/src/native_auth/controllers/sign_in/MSALNativeAuthInternalSignInParameters.swift
@@ -1,4 +1,3 @@
-//------------------------------------------------------------------------------
 //
 // Copyright (c) Microsoft Corporation.
 // All rights reserved.
@@ -22,22 +21,26 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-//
-//------------------------------------------------------------------------------
 
-#define MSAL_VER_HIGH       1
-#define MSAL_VER_LOW        7
-#define MSAL_VER_PATCH      0
+@_implementationOnly import MSAL_Private
 
-#define STR_HELPER(x) #x
-#define STR(x) STR_HELPER(x)
+class MSALNativeAuthInternalSignInParameters {
+    let username: String
+    let password: String?
+    let context: MSALNativeAuthRequestContext
+    let scopes: [String]?
+    let claimsRequestJson: String?
 
-// Framework versions only support high and low for the double value, sadly.
-#define MSAL_VERSION_STRING     STR(MSAL_VER_HIGH) "." STR(MSAL_VER_LOW) "." STR(MSAL_VER_PATCH)
-
-#import "IdentityCore_Internal.h"
-#import "MSIDLogger+Internal.h"
-#import "MSALError.h"
-#import "MSIDRequestContext.h"
-#import "MSALDefinitions.h"
-#import "MSALError.h"
+    init(
+        username: String,
+        password: String?,
+        context: MSALNativeAuthRequestContext,
+        scopes: [String]?,
+        claimsRequestJson: String?) {
+        self.username = username
+        self.password = password
+        self.context = context
+        self.scopes = scopes
+        self.claimsRequestJson = claimsRequestJson
+    }
+}