+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## TBD
+* Support sendable result #2518
+
+## [1.7.0]
+* Add support for claims request in native authentication signIn (#2496)
+* Move native auth public methods to parameter class (#2492)
+
 ## [1.6.3]
 * Merge 1.6.1-hotfix